requirements.txt
faker
python-docx
docx2pdf
pdf2image
pillow